backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
email-validator>=2.1.0
httpx
python-docx
PyPDF2
openai>=1.0.0
bcrypt==4.0.1